BookmarkFragment: Give outsiders the chance to decide which fragment is displayed when it becomes visible
BookmarkViewModel: Fix date issue when it is above 30 days

DisposableFragment: Use onViewCreated to add the scroll listeners as onActivityCreated was deprecated

HomeFragment: Display an error message when something goes wrong which we are informed using the boolean part of the Pair.
Use the progress bar in each RecyclerView adapter when refreshing and only display it if there was no item in the MangaListAdapter.

HomeViewModel: Changed to accept a Pair with the first being the index and the second a boolean that tells if successful.
A better way to show the error message
Fixed the repo call to download not executing

RecyclerViewAdapter: Can now show a text from the TextView in the layout file

SelectionFragment: removed onActivityCreated
Can now show a menu in the title toolbar

SingleNavigator: Removed unused instances

fragment_manga: Used string resources instead of literals

generic_loading_recycler_view: Added a minHeight and a TextView

Added a manga height dimen