Field renames in BookmarkFragment.java as well as a change to how the current item is selected.
Removed HistoryFragment.kt methods in BookmarkViewModel.kt, putting them in its own HistoryViewModel.kt.

Allow HistoryViewModel to be tied to the activity instead of the fragment to allow for better active updates from the DB

SearchViewFragment.java is a class that helps in handling search.

ViewPagerFragment.java renamed to BookmarkViewPagerFragment.java. Added a filter method to dynamically filter the mangas in it. Added a statement to know when the orientation of the device changes so as to change the amount of mangas displayed in a row.

Renamed setMenuResource to setContextMenuResource in SelectionFragment. Hides/Shows the BottomNavView when the Cab visibility changes.

Minor changes in InternetManager and Notifications

updateManga method in Repository now updates existing chapters and inserts new ones intead of simply inserting all of them which was causing few problems

Contemplating no longer using SingleNavigator to keep a single instance of the fragment. Added the settings menu to MainActivity

Replaced DisposableFragment.kt with DisposableViewModel.kt as the view model is usually the one making network calls. DisposableFragment.kt is therefore renamed to SwipeRefreshFragment.kt since thats what it does now

Fixed DateUtils only returning the current date