Moved Download class to the base model package.
Replaced DownloadManager.java with DownloadManager.kt

DownloadQueueFragment.java moved to the base adapter package

Added DownloadManga to properly display Downloads in the fragments

Added DownloadQueueMangaFragment to display the Mangas downloading and DownloadQueueFragment will display the chapters in a particular fragment that is downloading.

DownloadManager changes:
* Almost every method now comes with an explanation
* Downloads are now done by mangas
* A queue is made to manage the different Downloaders and is persisted with MangaPrefs

Downloader changes:
* Stopped creating an OkHttpClient instance itself instead receives one from the DownloadManager
* Addition of a concurrentDownloads field which will control how many downloads would be made at once and would be set by the user.
* Reset the progress of every Download to 0 before starting
* Replaced the doOnError, doOnNext, doOnComplete calls with their equivalents in the subscribe method as errors keep making the app crash.

DownloadNotifier changes:
* Now displays downloads as a group featuring every manga currently downloading
* Icons of the mangas are shown too

DownloadService changes:
* No longer manages Downloader. It simply informs the DownloadManager when downloads can start.

Created a MangaPrefs class to manage anything related to sharedpreferences

ChapterAdapter.java has been renamed to MangaInfoAdapter.java

Chapter changes:
* Added a new index
* Changed number to the actual number instead of a string.
* removed downloaded field, it can be checked from the download table
* offset field will be removed later

ElementExt changes:
* Statement to extract Alternative added
* A better means to identify latest chapters used
* Use regex to extract the chapter number instead of using the entire string, to extract a correct title which could also be in the chapter number

MangaFragment changes:
* RecyclerView issue solved by using just 1 adapter and putting the info as the first element and the list below it in the recycler view

MangaInfo now uses 2 different view model classes, the first for the manga information and the second for the chapters in the manga. ListAdapter was removed in order to handle the changes well

MangaViewModel now tries to determine if a manga should be updated instead of simply calling insert on it