PowerSpinner replaced with MaterialSpinner
Stopped using a TextView and ImageView for the header, instead it now uses just the text view as it has the capacity to also hold drawables

Split the addThumbnail method in BindingUtils so instead of creating the ImageTarget itself, it is passed to it.

HistoryFragment: Removed the default drawable of the header

DownloadMangaAdapter: Affected by BindingUtils changes

MangaSearch no longer uses a Builder for easy instantiation, it also now allows changes instead of simply being one sided.

Since all the information for a search is known before hand, creating all the involved views beforehand is easier than making a complex recycler view adapter implementation, so SearchDialogAdapter and search_edit_view has been scraped instead all the views are in SearchDialogFragment itself

MangaFragment and MangaInfoAdapter now helps to identify when completed downloads